requirements.txt
aiogram
pandas
openpyxl
sqlite3
python-dotenv
aiohttp
